desktop/requirements-pyqt.txt
PyQt5>=5.15.0
PyQtWebEngine>=5.15.0
python-dotenv>=1.0.0
